Used wp_remote_head to resolve mime-type.
According to the review feedback from plugins@wordpress.org, WordPress
comes with an extensive HTTP API that should be used instead of creating
your own curl calls. It’s both faster and more extensive. It’ll fall
back to curl if it has to, but it’ll use a lot of WordPress’ native
functionality first.

git-svn-id: https://plugins.svn.wordpress.org/external-media-without-import/trunk@1669263 b8457f37-d9ea-0310-8a92-e5e31aec5664